Code cleanup: refactor and delete obsolete classes
