--- app/save-and-restore/app/src/main/resources/org/phoebus/applications/saveandrestore/ui/snapshot/CompareSnapshotsTableView.fxml
@@ -1,69 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!--
-  ~ Copyright (C) 2020 European Spallation Source ERIC.
-  ~
-  ~  This program is free software; you can redistribute it and/or
-  ~  modify it under the terms of the GNU General Public License
-  ~  as published by the Free Software Foundation; either version 2
-  ~  of the License, or (at your option) any later version.
-  ~  
-  ~  This program is distributed in the hope that it will be useful,
-  ~  but WITHOUT ANY WARRANTY; without even the implied warranty of
-  ~  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-  ~  GNU General Public License for more details.
-  ~  
-  ~  You should have received a copy of the GNU General Public License
-  ~  along with this program; if not, write to the Free Software
-  ~  Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
-  ~  
-  -->
-
-<?import javafx.scene.control.*?>
-<?import javafx.scene.control.cell.*?>
-<?import javafx.scene.layout.*?>
-<?import org.phoebus.applications.saveandrestore.ui.snapshot.*?>
-
-<TableView fx:id="snapshotTableView" editable="true" maxWidth="1.7976931348623157E308" HBox.hgrow="ALWAYS" xmlns="http://javafx.com/javafx/17.0.2-ea" xmlns:fx="http://javafx.com/fxml/1"
-           fx:controller="org.phoebus.applications.saveandrestore.ui.snapshot.CompareSnapshotsTableViewController">
-    <columns>
-        <SelectionTableColumn fx:id="selectedColumn" minWidth="30" prefWidth="30" resizable="false" tooltip="%includeThisPV" visible="false">
-            <cellValueFactory>
-                <PropertyValueFactory property="selected" />
-            </cellValueFactory>
-        </SelectionTableColumn>
-
-        <TooltipTableColumn fx:id="idColumn" labelText="#" resizable="false" tooltip="%toolTipTableColumIndex" />
-
-        <TooltipTableColumn labelText="%tableColumnPVName" minWidth="100.0" prefWidth="100.0" styleClass="snapshot-table-left-aligned" tooltip="%toolTipTableColumnPVName">
-            <cellValueFactory>
-                <PropertyValueFactory property="pvName" />
-            </cellValueFactory>
-        </TooltipTableColumn>
-
-        <DividerTableColumn/>
-
-        <TableColumn styleClass="toplevel" text="%storedValues">
-            <columns>
-                <TableColumn fx:id="baseSnapshotName" styleClass="second-level">
-                    <columns>
-                        <TooltipTableColumn fx:id="storedValueColumn" labelText="%tableColumnSetpoint" minWidth="100" prefWidth="100" styleClass="snapshot-table-left-aligned" tooltip="%toolTipTableColumnSetpointPVValue">
-                            <cellValueFactory>
-                                <PropertyValueFactory property="storedValue" />
-                            </cellValueFactory>
-                        </TooltipTableColumn>
-                        <TableColumn fx:id="deltaColumn" minWidth="100" prefWidth="100" styleClass="snapshot-table-centered" text="%tableColumnDeltaValue" />
-                    </columns>
-                </TableColumn>
-                <DividerTableColumn/>
-            </columns>
-        </TableColumn>
-
-        <TooltipTableColumn fx:id="liveValueColumn" editable="false" labelText="%tableColumnLiveSetpoint" maxWidth="1.7976931348623157E308" minWidth="100" sortable="false" styleClass="snapshot-table-left-aligned" tooltip="%currentPVValue">
-            <cellValueFactory>
-                <PropertyValueFactory property="liveValue" />
-            </cellValueFactory>
-        </TooltipTableColumn>
-
-    </columns>
-</TableView>
